+++ src/App.js
@@ -1,45 +1,18 @@
-import React, {useState, useEffect} from "react"
-import OneMission from "./OneMission"
+import React from "react"
+import { BrowserRouter, Route, Switch } from "react-router-dom"
 
-const AllInfo = props => {
-  const [info, setInfo] = useState ([])
+import IndexContainer from "./IndexContainer"
+import ShowContainer from "./ShowContainer"
 
-  useEffect (() => {
-  fetch('https://api.spacexdata.com/v3/launches')
-  .then(response =>{
-    if(response.ok){
-      return response;
-      } else {
-        let errorMessage = `${response.status}
-        (${response.statusText})`,
-        error = new Error(errorMessage);
-        throw(error)
-      }
-    })
-  .then(response => response.json())
-  .then(responseFromServer =>{
-    setInfo(responseFromServer)
-    })
-  .catch(error => console.error(`Error in fetch
-    ${error.message}`))
-  }, [])
-
-  const eachMissionInfo = info.map(info =>{
-    return (
-      <OneMission
-        key={info.mission_name}
-        missionName={info.mission_name}
-        flightNumber={info.flight_number}
-      />
-    )
-  })
-
+const App = props => {
   return (
-    <div>
-      <h1> SpaceX API : </h1>
-        {eachMissionInfo}
-    </div>
+    <BrowserRouter>
+      <Switch>
+        <Route exact path="/" component={IndexContainer}/>
+        <Route exact path="/launches/id" component={ShowContainer}/>
+      </Switch>
+    </BrowserRouter>
   )
 }
 
-export default AllInfo;
+export default App
