+++ src/ShowContainer.js
@@ -1,0 +1,10 @@
+import React from "react"
+
+const oneMissionInfo = props => {
+  return(
+      <p>hey
+      </p>
+  )
+}
+
+export default oneMissionInfo
